refactor(ui): extract ImageAttachments component from MessageInput
Pull paste/file-picker/thumbnail logic into reusable useImageAttachments hook. Adds 10MB size cap and exposes clean interface for attachment management.
